requirements.txt
Pillow==8.2.0
django-cors-headers==3.7.0
djangorestframework==3.12.4